backend/requirements.txt
openai-agents>=0.6.8
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
websockets>=12.0
